requirements.txt
discord.py
requests
aiohttp
pymongo
dnspython